- Add transition to character portraits - Update code
