+++ MarvelCharsApp/MarvelCharsApp/ViewModels/HomeViewModel.swift
@@ -9,7 +9,7 @@ import Foundation
 
 protocol HomeViewModelDelegate {
   func onFetchDataSuccessfully(categories: [CategoryModel])
-  func onFetchHeroes(heroes: CategoryModel)
+  
 }
 
 class HomeViewModel {
@@ -18,11 +18,6 @@ class HomeViewModel {
   var delegate: HomeViewModelDelegate?
   var categories: [CategoryModel] = []
   var categoriesNames: [String] = []
-  var heroes: [CharacterModel] = []
-  var villains: [CharacterModel] = []
-  var antiHeroes: [CharacterModel] = []
-  var aliens: [CharacterModel] = []
-  var humans: [CharacterModel] = []
   
   init(){
     self.manager.delegate = self
@@ -45,11 +40,6 @@ class HomeViewModel {
 extension HomeViewModel: MarvelAPIDelegate {
   
   func didFetchData(categories: [CategoryModel]) {
-    self.heroes = categories[0].characters
-    self.villains = categories[1].characters
-    self.antiHeroes = categories[2].characters
-    self.aliens = categories[3].characters
-    self.humans = categories[4].characters
     self.categories = categories
     self.categoriesNames = self.categories.map{ category in category.category}
     self.delegate?.onFetchDataSuccessfully(categories: categories)
