+++ MarvelCharsApp/MarvelCharsApp/Views/Cells/CategoryRowCell.xib
@@ -33,7 +33,7 @@
                     </button>
                     <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="P2W-e2-t3Q">
                         <rect key="frame" x="0.0" y="47" width="320" height="270"/>
-                        <color key="backgroundColor" red="0.89000000000000001" green="0.77589743590000004" blue="0.87693545110000004" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
                             <constraint firstAttribute="height" constant="270" id="BcP-Z6-GFa"/>
                         </constraints>
@@ -68,6 +68,9 @@
         <namedColor name="SeeAllButtonLabelColor">
             <color red="0.32899999618530273" green="0.32400000095367432" blue="0.4779999852180481" alpha="0.33000001311302185" colorSpace="custom" customColorSpace="sRGB"/>
         </namedColor>
+        <systemColor name="systemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
         <systemColor name="systemRedColor">
             <color red="1" green="0.23137254901960785" blue="0.18823529411764706" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </systemColor>
