+++ MarvelCharsApp/MarvelCharsApp/Views/Cells/CharacterPortraitCell.swift
@@ -19,13 +19,3 @@ class CharacterPortraitCell: UICollectionViewCell {
   }
   
 }
-
-extension UIImage{
-  var roundedImage: UIImage{
-    let rect = CGRect(origin: CGPoint(x: 0, y: 0), size: self.size)
-    UIGraphicsBeginImageContextWithOptions(self.size, false, 1)
-    UIBezierPath(roundedRect: rect, cornerRadius: 50).addClip()
-    self.draw(in: rect)
-    return UIGraphicsGetImageFromCurrentImageContext()!
-  }
-}
